requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
google-genai>=0.3.0